Separated io-elements from main build.
