+++ src/core/bindings.js
