+++ src/core/node.js
