+++ src/core/properties.js
