+++ src/elements/collapsable.js
@@ -1,4 +1,4 @@
-import {html, IoElement} from "./core/element.js";
+import {html, IoElement} from "../core/element.js";
 
 export class IoCollapsable extends IoElement {
   static get Style() {
