+++ src/io-elements.js
@@ -1,0 +1,48 @@
+/** @license
+ * The MIT License
+ *
+ * Copyright © 2019 Aleksandar (Aki) Rodić
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+export {IoBoolean} from "./elements/boolean.js";
+export {IoBreadcrumbs} from "./elements/breadcrumbs.js";
+export {IoButton} from "./elements/button.js";
+export {IoCollapsable} from "./elements/collapsable.js";
+export {IoInspector} from "./elements/inspector.js";
+export {IoGl} from "./elements/gl.js";
+export {IoItem} from "./elements/item.js";
+export {IoLayout} from "./elements/layout.js";
+export {IoMenuItem} from "./elements/menu-item.js";
+export {IoMenuLayer} from "./elements/menu-layer.js";
+export {IoMenuOptions} from "./elements/menu-options.js";
+export {IoMenu} from "./elements/menu.js";
+export {IoNumber} from "./elements/number.js";
+export {IoObject} from "./elements/object.js";
+export {IoOption} from "./elements/option.js";
+export {IoProperties} from "./elements/properties.js";
+export {IoSelectorSidebar} from "./elements/selector-sidebar.js";
+export {IoSelectorTabs} from "./elements/selector-tabs.js";
+export {IoSelector} from "./elements/selector.js";
+export {IoSlider} from "./elements/slider.js";
+export {IoString} from "./elements/string.js";
+export {IoSidebar} from "./elements/sidebar.js";
+export {IoSwitch} from "./elements/switch.js";
+export {IoTheme} from "./elements/theme.js";
