+++ src/io.js
@@ -22,32 +22,7 @@
  * THE SOFTWARE.
  */
 
-export {IoNodeMixin, IoNode} from "./io/core/node.js";
-export {IoElement, html} from "./io/core/element.js";
-export {Binding, NodeBindings} from "./io/core/bindings.js";
-export {IoStorage, storageNodes} from "./io/core/storage.js";
-
-export {IoBoolean} from "./io/boolean.js";
-export {IoBreadcrumbs} from "./io/breadcrumbs.js";
-export {IoButton} from "./io/button.js";
-export {IoCollapsable} from "./io/collapsable.js";
-export {IoInspector} from "./io/inspector.js";
-export {IoGl} from "./io/gl.js";
-export {IoItem} from "./io/item.js";
-export {IoLayout} from "./io/layout.js";
-export {IoMenuItem} from "./io/menu-item.js";
-export {IoMenuLayer} from "./io/menu-layer.js";
-export {IoMenuOptions} from "./io/menu-options.js";
-export {IoMenu} from "./io/menu.js";
-export {IoNumber} from "./io/number.js";
-export {IoObject} from "./io/object.js";
-export {IoOption} from "./io/option.js";
-export {IoProperties} from "./io/properties.js";
-export {IoSelectorSidebar} from "./io/selector-sidebar.js";
-export {IoSelectorTabs} from "./io/selector-tabs.js";
-export {IoSelector} from "./io/selector.js";
-export {IoSlider} from "./io/slider.js";
-export {IoString} from "./io/string.js";
-export {IoSidebar} from "./io/sidebar.js";
-export {IoSwitch} from "./io/switch.js";
-export {IoTheme} from "./io/theme.js";
+export {IoNodeMixin, IoNode} from "./core/node.js";
+export {IoElement, html} from "./core/element.js";
+export {Binding, NodeBindings} from "./core/bindings.js";
+export {IoStorage, storageNodes} from "./core/storage.js";
